fixes #11
