+++ Core/GenerateJson.cs
@@ -1,0 +1,7 @@
+﻿namespace lvfucs.Core
+{
+	public class GenerateJson
+	{
+	}
+}
+
